moved erroreventdefinition
